color conversion chunks of chunks
The idea is, now par1 processes each pixel which consists of color channels in parallel. But I want to somehow make it such that each of these pixels are also worked on simultaneously in threads. Basically, chunks of chunks of pixels rather than chunk of pixel channels.

Problems: For some reason, par2 fails at img2. It processes correctly for img1 and img3 but not the one in between.